Update README.md
Thought to let you know that I made some tests on this and I found workarounds, beacuse this looked to fail. So, found the correct way of building this. All it needs are some colons that are missing in the main.js to fix as I did and just for the build.sh to work correct. electron installation wasn't smooth and I found a workaround for it to work as a global installa. Another good thing is to make a custom menu in electron main.js and explain to the user if they would like to change their info on the menu within the main.js file to rerflect their own info.